tests/wpt/web-platform-tests/layout-instability/shift-with-overflow-status-change.html

<!DOCTYPE html>
<title>Layout Instability: change under overflow clipping container causing shift and overflow status change at the same time</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<div style="height: 100px"></div>
<div style="overflow: auto; width: 400px; height: 400px">
  <div id="resized" style="width: 600px; height: 100px; background: gray"></div>
  <div id="shifted" style="width: 300px; height: 100px; background: blue"></div>
</div>
<script>
promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  resized.style.width = '200px';
  resized.style.height = '200px';

  const expectedScore = computeExpectedScore(300 * (100 + 100), 100);

  assert_equals(watcher.score, 0);
  await watcher.promise;
  assert_equals(watcher.score, expectedScore);

  resized.style.width = '600px';
  resized.style.height = '100px';
  await watcher.promise;
  assert_equals(watcher.score, expectedScore * 2);
}, 'Change under overflow clipping container causing shift and overflow status change at the same time');
</script>
